Add CreateCourseHandler
